+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Evolutions de l'extension
 
+## [1.6.0] 17/05/2025
+
+- iconv-lite en modules et readFileSync writeFileSync avec encodage iconv
+- fs et path en modules
+
+## [1.5.0] 17/05/2025
+
+- Detection du changement d'une macro et réinstallation avant appel.
+
 ## [1.3.0] 17/05/2025
 
 - Exec Commande ajouté.